+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 - renaming InformationDisplay to DataDisplay
 - [MapControls] - from presets to Forms/Actions/MapControl/presets/MapControls
 - [IconMarkers] - from presets to Geospatial/IconMarker/presets/IconMarkers
+- [Pagination] - from DataDiplay to Navigation
 
 ## [v2.131.0] - 09-10-2025